feat (#64): Validação dos dados de cadastro e conexão com frontend para receber e enviar dados do back.

+++ profissionais/src/main/java/br/com/chronos/profissionais/api/dto/ProjetoVinculadoResposta.java
@@ -1,11 +1,9 @@
 package br.com.chronos.profissionais.api.dto;
 
-import java.math.BigDecimal;
-
 public record ProjetoVinculadoResposta(
-        Integer projetoId,
+        int projetoId,
         String nomeProjeto,
         String codigoProjeto,
-        BigDecimal valorHora
+        double valorHora
 ) {
 }
